backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OpenAI
openai>=1.12.0

# Vector Database
chromadb==1.3.4

# LangChain (for easier integration)
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# PDF Processing
PyMuPDF==1.23.20
pdfplumber==0.10.3
Pillow==10.2.0

# Utilities
python-multipart==0.0.6
httpx==0.28.0
aiofiles==23.2.1
tqdm==4.67.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

